Move upload out of Menu and into ProjectManage
